move tests to test_checkpoint
